Move test-inputs.ts into src/ to check types of fixtures
